feat : arrange schema
전반적인 DB 스키마 정리
mongoose code 삭제
