+++ data/mock.js
@@ -1,2038 +1,221 @@
-export const productsData = [
-  {
-    _id: "9F3F1E600554D3E69F1F6133",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [
-      "66b5ccad2cf6ddb53d85e3a6",
-      "66b5ccad2cf6ddb53d85e3a5",
-      "66b5ccad2cf6ddb53d85e3a4",
-    ],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6132",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [
-      "66b5ccad2cf6ddb53d85e3a6",
-      "66b5ccad2cf6ddb53d85e3a5",
-      "66b5ccad2cf6ddb53d85e3a4",
-    ],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6131",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [
-      "66b5ccad2cf6ddb53d85e3a6",
-      "66b5ccad2cf6ddb53d85e3a5",
-      "66b5ccad2cf6ddb53d85e3a4",
-    ],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6130",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: ["66b5ccad2cf6ddb53d85e3a6", "66b5ccad2cf6ddb53d85e3a5"],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6129",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: ["66b5ccad2cf6ddb53d85e3a6", "66b5ccad2cf6ddb53d85e3a5"],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6128",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: ["66b5ccad2cf6ddb53d85e3a6", "66b5ccad2cf6ddb53d85e3a5"],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6127",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: ["66b5ccad2cf6ddb53d85e3a6"],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6126",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6125",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6124",
-    name: "맥북 프로",
-    description: "애플 맥북 프로 13인치",
-    price: 1500000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6123",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6122",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6121",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6120",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6119",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6118",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6117",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6116",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6115",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6114",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6113",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6112",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6111",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6110",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6109",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6108",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6107",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6106",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6105",
-    name: "맥북 프로",
-    description: "애플 맥북 프로 13인치",
-    price: 1500000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6104",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6103",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6102",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6101",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6100",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6099",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6098",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6097",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6096",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6095",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6094",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6093",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6092",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6091",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6090",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6089",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6088",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6087",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6086",
-    name: "맥북 프로",
-    description: "애플 맥북 프로 13인치",
-    price: 1500000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6085",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6084",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6083",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6082",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6081",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6080",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6079",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6078",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6077",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6076",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6075",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6074",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6073",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6072",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6071",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6070",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6069",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6068",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6067",
-    name: "맥북 프로",
-    description: "애플 맥북 프로 13인치",
-    price: 1500000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6066",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6065",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6064",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6063",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6062",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6061",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6060",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6059",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6058",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6057",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6056",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6055",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6054",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6053",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6052",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6051",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6050",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6049",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6048",
-    name: "맥북 프로",
-    description: "애플 맥북 프로 13인치",
-    price: 1500000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6047",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6046",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6045",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6044",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6043",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6042",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6041",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6040",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6039",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6038",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6037",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6036",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6035",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6034",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6033",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6032",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6031",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6030",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6029",
-    name: "맥북 프로",
-    description: "애플 맥북 프로 13인치",
-    price: 1500000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6028",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6027",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6026",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6025",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6024",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6023",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6022",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6021",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6020",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6019",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6018",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6017",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6016",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6015",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6014",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6013",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6012",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6011",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6010",
-    name: "맥북 프로",
-    description: "애플 맥북 프로 13인치",
-    price: 1500000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9",
-    name: "스니커즈",
-    description: "편안한 스니커즈",
-    price: 100000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991853452/5389615.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6008",
-    name: "갤럭시 탭 S7",
-    description: "삼성 갤럭시 탭 S7",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991844193/5146532.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6007",
-    name: "보스 헤드폰",
-    description: "보스 노이즈 캔슬링 헤드폰 700",
-    price: 350000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991827255/3514562.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6006",
-    name: "사무용 의자",
-    description: "편안한 사무용 의자",
-    price: 120000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991812368/2313561.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6005",
-    name: "레노버 노트북",
-    description: "레노버 아이디어패드 5",
-    price: 800000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991798558/321351.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6004",
-    name: "갤럭시 S21",
-    description: "삼성 갤럭시 S21",
-    price: 600000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991786504/31563.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6003",
-    name: "아이패드",
-    description: "애플 아이패드 10.2인치",
-    price: 450000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991756711/21353.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6002",
-    name: "퀸사이즈 침대",
-    description: "퀸사이즈 침대 프레임",
-    price: 500000,
-    tags: ["가구"],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991744735/2113.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    _id: "9F3F1E600554D3E69F1F6001",
-    name: "판다인형",
-    description: "귀여운 판다인형입니다",
-    price: 30000,
-    tags: [],
-    images: [
-      "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Sprint_Mission/user/3/1721991725132/233.png",
-    ],
-    ownerId: "1",
-    favorite_user_id: [],
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+export const User = [
+  {
+    id: "123e4567-e89b-12d3-a456-426614174001",
+    name: "이진우",
+    nickname: "몽유랑",
+    password: "$2a$10$fL8Q6iW1xj4oRk5lS9T0mUqJtVwZ7xNz3eYpXrDyCvB",
+    email: "jinwoo@example.com",
+    image: "https://ifh.cc/g/cSYLVn.png",
+    createdAt: "2022-01-01T00:00:00.000Z",
+    updatedAt: "2022-01-02T12:34:56.789Z",
+  },
+  {
+    id: "7890abc1-def2-g345-h6789012345678",
+    name: "안두인",
+    nickname: "Allience",
+    password: "$2a$10$fL8Q6iW1xj4oRk5lS9T0mUqJtVwZ7xNz3eYpXrDyCvB",
+    email: "and@example.com",
+    image: "https://ifh.cc/g/ojY5Vr.jpg",
+    createdAt: "2022-01-01T00:00:00.000Z",
+    updatedAt: "2022-01-02T12:34:56.789Z",
+  },
+  {
+    id: "345e6789-cdef-ghij-klmnop-12345678901",
+    name: "스랄",
+    nickname: "Horde",
+    password: "$2a$10$fL8Q6iW1xj4oRk5lS9T0mUqJtVwZ7xNz3eYpXrDyCvB",
+    email: "thr@example.com",
+    image: "https://ifh.cc/g/25YQq2.jpg",
+    createdAt: "2022-01-01T00:00:00.000Z",
+    updatedAt: "2022-01-02T12:34:56.789Z",
+  },
+  {
+    id: "345e6789-cdef-ghij-klmnop-12345678902",
+    name: "실리안",
+    nickname: "루테란 국왕",
+    password: "$2a$10$fL8Q6iW1xj4oRk5lS9T0mUqJtVwZ7xNz3eYpXrDyCvB",
+    email: "sil@example.com",
+    image: "https://ifh.cc/g/1btdoy.png",
+    createdAt: "2022-01-01T00:00:00.000Z",
+    updatedAt: "2022-01-02T12:34:56.789Z",
+  },
+  {
+    id: "345e6789-cdef-ghij-klmnop-12345678903",
+    name: "아만",
+    nickname: "로스트아크",
+    password: "$2a$10$fL8Q6iW1xj4oRk5lS9T0mUqJtVwZ7xNz3eYpXrDyCvB",
+    email: "arm@example.com",
+    image: "https://ifh.cc/g/ANb6SK.png",
+    createdAt: "2022-01-01T00:00:00.000Z",
+    updatedAt: "2022-01-02T12:34:56.789Z",
+  },
+  {
+    id: "345e6789-cdef-ghij-klmnop-12345678904",
+    name: "니나브",
+    nickname: "에스더",
+    password: "$2a$10$fL8Q6iW1xj4oRk5lS9T0mUqJtVwZ7xNz3eYpXrDyCvB",
+    email: "nin@example.com",
+    image: "https://ifh.cc/g/DfVJYO.png",
+    createdAt: "2022-01-01T00:00:00.000Z",
+    updatedAt: "2022-01-02T12:34:56.789Z",
   },
 ];
 
-export const userData = [
+export const Product = [
   {
-    name: "사용일",
-    nickname: "user1",
-    password: "12345678",
-    email: "user1@codeit.com",
-    image:
-      "https://slp-statics.astockcdn.net/static_assets/staging/23spring/kr/home/curated-collections/card-1.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "7890abc1-def2-g345-h6789012345678",
+    name: "PS5 팝니다",
+    description: "현생이 바빠서 기기 처분합니다",
+    price: 350000,
+    userId: "123e4567-e89b-12d3-a456-426614174001",
+    createdAt: "2022-01-01T00:00:00.000Z",
+    updatedAt: "2022-01-02T12:34:56.789Z",
   },
   {
-    name: "사용이",
-    nickname: "user2",
-    password: "12345678",
-    email: "user2@codeit.com",
-    image:
-      "https://slp-statics.astockcdn.net/static_assets/staging/23spring/kr/home/curated-collections/card-5.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "345e6789-cdef-ghij-klmnop-12345678901",
+    name: "PS 타이틀 팝니다",
+    description: "현생이 너무 바빠 정리합니다",
+    price: 240000,
+    userId: "123e4567-e89b-12d3-a456-426614174001",
+    createdAt: "2022-01-01T00:00:00.000Z",
+    updatedAt: "2022-01-02T12:34:56.789Z",
   },
   {
-    name: "사용삼",
-    nickname: "user3",
-    password: "12345678",
-    email: "user3@codeit.com",
-    image:
-      "https://t4.ftcdn.net/jpg/06/66/39/51/240_F_666395131_xcH5fxuuGTrC5kQxuYDgbFBrQDeNBCWJ.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "345e6789-cdef-ghij-klmnop-12345678902",
+    name: "스틈윈드 특산 포도주 팝니다",
+    description: "내부 전쟁 준비로 비품 확인 중 찾은 귀한 포도주 팝니다",
+    price: 2000000,
+    userId: "7890abc1-def2-g345-h6789012345678",
+    createdAt: "2022-01-01T00:00:00.000Z",
+    updatedAt: "2022-01-02T12:34:56.789Z",
   },
   {
-    name: "사용사",
-    nickname: "user4",
-    password: "12345678",
-    email: "user4@codeit.com",
-    image:
-      "https://t4.ftcdn.net/jpg/05/20/29/39/240_F_520293939_e0xAeUAPmdy1cKmuQlr2JjH9yZsorckz.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "345e6789-cdef-ghij-klmnop-12345678903",
+    name: "오그리마 특산 텐트 팝니다",
+    description: "어디서든 편하게 설치 가능한 텐트 팝니다. 오그리마 산이예요~",
+    price: 500000,
+    userId: "345e6789-cdef-ghij-klmnop-12345678901",
+    createdAt: "2022-01-01T00:00:00.000Z",
+    updatedAt: "2022-01-02T12:34:56.789Z",
+  },
+];
+
+export const ProductTags = [];
+
+export const ProductImages = [];
+
+export const FavoriteProduct = [];
+
+export const Post = [
+  {
+    id: "123e4567-e89b-12d3-a456-426614174004",
+    userId: "123e4567-e89b-12d3-a456-426614174001",
+    title: "여기가 자유 게시판인가요?",
+    content: "첫 글 성공",
+    favoriteCount: 3,
+    createdAt: "2022-01-01T00:00:00.000Z",
+    updatedAt: "2022-01-02T12:34:56.789Z",
   },
   {
-    name: "사용오",
-    nickname: "user5",
-    password: "12345678",
-    email: "user5@codeit.com",
-    image:
-      "https://t3.ftcdn.net/jpg/06/11/91/88/240_F_611918802_H1NQufUojOlkSZeraPui4YKilsj4PJVI.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "123e4567-e89b-12d3-a456-426614174005",
+    userId: "7890abc1-def2-g345-h6789012345678",
+    title: "요즘 와우 재미있나요?",
+    content: "평가도 좋던데",
+    favoriteCount: 1,
+    createdAt: "2022-02-03T15:16:17.890Z",
+    updatedAt: "2022-02-04T18:19:20.123Z",
   },
   {
-    name: "사용육",
-    nickname: "user6",
-    password: "12345678",
-    email: "user6@codeit.com",
-    image:
-      "https://t4.ftcdn.net/jpg/01/08/15/27/240_F_108152713_Bo237JlV0olJxajFORewebhiZ0hcm0Mo.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "123e4567-e89b-12d3-a456-426614174006",
+    userId: "345e6789-cdef-ghij-klmnop-12345678903",
+    title: "요즘 로아 재미있나요?",
+    content: "초각성 나왔습니다",
+    favoriteCount: 1,
+    createdAt: "2024-07-03T15:16:17.890Z",
+    updatedAt: "2024-07-04T18:19:20.123Z",
+  },
+];
+
+export const FavoritePost = [
+  {
+    id: "7890a001-def2-g345-h6789012345682",
+    userId: "7890abc1-def2-g345-h6789012345678",
+    postId: "123e4567-e89b-12d3-a456-426614174004",
+    createdAt: "2024-02-03T15:16:17.890Z",
+    updatedAt: "2024-02-03T15:16:17.890Z",
   },
   {
-    name: "사용칠",
-    nickname: "user7",
-    password: "12345678",
-    email: "user7@codeit.com",
-    image:
-      "https://t3.ftcdn.net/jpg/02/88/21/50/240_F_288215005_B7LfitWNnvCNeV9FJ8JU9m0Z16Cqi17k.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "7890a001-def2-g345-h6789012345683",
+    userId: "345e6789-cdef-ghij-klmnop-12345678903",
+    postId: "123e4567-e89b-12d3-a456-426614174004",
+    createdAt: "2024-02-03T15:16:17.890Z",
+    updatedAt: "2024-02-03T15:16:17.890Z",
   },
   {
-    name: "사용팔",
-    nickname: "user8",
-    password: "12345678",
-    email: "user8@codeit.com",
-    image:
-      "https://t4.ftcdn.net/jpg/05/74/18/55/240_F_574185531_6whwbrEXvISh4qyug5KtZIaL3673YYJP.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "7890a001-def2-g345-h6789012345684",
+    userId: "345e6789-cdef-ghij-klmnop-12345678904",
+    postId: "123e4567-e89b-12d3-a456-426614174004",
+    createdAt: "2024-02-03T15:16:17.890Z",
+    updatedAt: "2024-02-03T15:16:17.890Z",
   },
   {
-    name: "사용구",
-    nickname: "user9",
-    password: "12345678",
-    email: "user9@codeit.com",
-    image:
-      "https://t4.ftcdn.net/jpg/06/24/13/83/240_F_624138301_jALchb7zdhEmXjAbjWqklrZt4EqdbfCQ.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "7890a001-def2-g345-h6789012345685",
+    userId: "123e4567-e89b-12d3-a456-426614174001",
+    postId: "123e4567-e89b-12d3-a456-426614174005",
+    createdAt: "2024-02-03T15:16:17.890Z",
+    updatedAt: "2024-02-03T15:16:17.890Z",
   },
   {
-    name: "사용십",
-    nickname: "user10",
-    password: "12345678",
-    email: "user10@codeit.com",
-    image:
-      "https://t3.ftcdn.net/jpg/06/03/18/50/240_F_603185033_AqvhJZB4lfcWAMqVTUnrJaYGDZMJO0oB.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "7890a001-def2-g345-h6789012345686",
+    userId: "123e4567-e89b-12d3-a456-426614174001",
+    postId: "123e4567-e89b-12d3-a456-426614174006",
+    createdAt: "2024-02-03T15:16:17.890Z",
+    updatedAt: "2024-02-03T15:16:17.890Z",
+  },
+];
+
+export const PostComment = [
+  {
+    id: "123e4567-e89b-12d3-a456-426614174005",
+    userId: "123e4567-e89b-12d3-a456-426614174001",
+    postId: "123e4567-e89b-12d3-a456-426614174004",
+    content: "성공",
+    createdAt: "2022-01-05T21:22:23.456Z",
+    updatedAt: "2022-01-06T00:03:04.789Z",
   },
   {
-    name: "유재석",
-    nickname: "메뚜기",
-    password: "12345678",
-    email: "user11@codeit.com",
-    image:
-      "https://img.seoul.co.kr/img/upload/2023/02/05/SSC_20230205232436_O2.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "7890abc1-def2-g345-h6789012345682",
+    userId: "123e4567-e89b-12d3-a456-426614174001",
+    postId: "123e4567-e89b-12d3-a456-426614174005",
+    content: "지인들이 다시 시작해서 관심은 가는데 현생이...",
+    createdAt: "2022-02-07T03:04:05.123Z",
+    updatedAt: "2022-02-08T06:07:08.456Z",
   },
   {
-    name: "박명수",
-    nickname: "활명수",
-    password: "12345678",
-    email: "user12@codeit.com",
-    image: "https://i3n.news1.kr/system/photos/2022/4/7/5306119/high.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "7890abc1-def2-g345-h6789012345683",
+    userId: "123e4567-e89b-12d3-a456-426614174001",
+    postId: "123e4567-e89b-12d3-a456-426614174006",
+    content: "출석 보상도 겨우 받고 있습니다 ㅠㅠ",
+    createdAt: "2022-02-07T03:04:05.123Z",
+    updatedAt: "2022-02-08T06:07:08.456Z",
+  },
+];
+
+export const ProductComment = [
+  {
+    id: "123e4567-e89b-12d3-a456-426614174005",
+    userId: "123e4567-e89b-12d3-a456-426614174001",
+    productId: "345e6789-cdef-ghij-klmnop-12345678901",
+    content: "This is my first product comment!",
+    createdAt: "2022-01-05T21:22:23.456Z",
+    updatedAt: "2022-01-06T00:03:04.789Z",
   },
   {
-    name: "정형돈",
-    nickname: "미존개오",
-    password: "12345678",
-    email: "user13@codeit.com",
-    image:
-      "https://img9.yna.co.kr/photo/cms/2020/11/06/50/PCM20201106000150990_P2.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    name: "하하",
-    nickname: "꼬마",
-    password: "12345678",
-    email: "user14@codeit.com",
-    image:
-      "https://pimg.mk.co.kr/meet/neds/2021/01/image_readtop_2021_2207_16095448504493586.jpeg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    name: "노홍철",
-    nickname: "돌아이",
-    password: "12345678",
-    email: "user15@codeit.com",
-    image:
-      "https://thumbnews.nateimg.co.kr/view610///news.nateimg.co.kr/orgImg/xs/2024/07/07/1720343343721523.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
-  },
-  {
-    name: "이진우",
-    nickname: "설랑",
-    password: "12345678",
-    email: "user99@codeit.com",
-    image:
-      "https://t3.ftcdn.net/jpg/07/59/85/60/240_F_759856080_xuXtGdTBtKKZCqyfsn38ypm1SFVczDIU.jpg",
-    createdAt: "2024-07-29T05:45:03.249Z",
-    updatedAt: "2024-07-29T05:45:03.249Z",
+    id: "7890abc1-def2-g345-h6789012345682",
+    userId: "7890abc1-def2-g345-h6789012345678",
+    productId: "345e6789-cdef-ghij-klmnop-12345678901",
+    content: "This is my second product comment!",
+    createdAt: "2022-02-07T03:04:05.123Z",
+    updatedAt: "2022-02-08T06:07:08.456Z",
   },
 ];
