+++ http/postgreSQL/article.http
@@ -27,10 +27,12 @@ Authorization : 123e4567-e89b-12d3-a456-426614174001
 ###
 GET http://localhost:3000/article?pageSize=3&page=1
 Content-Type : application/json
-Authorization : 7890abc1-def2-g345-h6789012345678
+Authorization : 123e4567-e89b-12d3-a456-426614174001
 
 ### 
 GET https://sprint-be.onrender.com/article
+Content-Type : application/json
+Authorization : 123e4567-e89b-12d3-a456-426614174001
 
 ###
 GET http://localhost:3000/article/123e4567-e89b-12d3-a456-426614174004
